% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimbalance.R
\name{bimbalance}
\alias{bimbalance}
\title{bimbalance - balanced item analysis}
\usage{
bimbalance(dat = NULL, sidevar = "sideA")
}
\arguments{
\item{dat}{raw data (e.g., from the \code{bimload} function)}

\item{sidevar}{the actual column name of the side variable in the raw data
(must be coded "right" or "left")}
}
\value{
data.frame R/L counts and the RL_ratio
}
\description{
The \code{bimbalance} function checks the imbalances of all present item
combinations in the data in terms of side (R: right, L:left). It counts the
cases on each side and also calculates a ratio (RL_ratio = R/L).
}
