% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimeval.R
\name{bimeval}
\alias{bimeval}
\title{Worth Values Evaluation Function}
\usage{
bimeval(
  ydata = NULL,
  worth = NULL,
  GT = NULL,
  simOpt = NULL,
  coverage = 0.8,
  showPlot = TRUE,
  title = "Consensus Analysis",
  subtitle = NULL
)
}
\arguments{
\item{ydata}{curated data.frame from the preprocessing function.}

\item{worth}{the worth matrix from the bimworth function. Note, that this can be a list when the intransitivity is calculated!}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new}

\item{coverage}{the threshold for the ratio of tested subjects per total subjects (default=0.8)}

\item{showPlot}{plot the worth plot as a bubble plot with uncertainty errors}

\item{title}{plot title (default: "Consensus Analysis")}

\item{subtitle}{plot subtitle (if NULL, a default will be used)}
}
\value{
either a data.frame with the consensus errors and/or a plot object
}
\description{
The \code{bimeval} will evaluate the worth values and will give an estimate
on the error of uncertainty in the unsimulated data.
}
