% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimintrans.R
\name{bimintrans}
\alias{bimintrans}
\title{Intransitivity Calculation}
\usage{
bimintrans(
  dat = NULL,
  idcolumn = "subjectID",
  I2 = "optionA",
  I1 = "optionB",
  response = "result"
)
}
\arguments{
\item{dat}{uses sorted pairings of input data from a preference test.}

\item{idcolumn}{variable name for the ID column (default is subjectID)}

\item{I2}{item 2 - optionA}

\item{I1}{item 1 - optionB}

\item{response}{response column}
}
\value{
list with the total number of intransitivity counts and number of triplets
}
\description{
The \code{bimintrans} calculates the intransitivity of items.
}
