% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimload.R
\name{bimload}
\alias{bimload}
\title{Import Function}
\usage{
bimload(filename)
}
\arguments{
\item{filename}{path to the raw data in *.txt format}
}
\value{
data.frame with the raw data
}
\description{
The \code{bimload} function loads the raw data as a data frame.
The function has several control and helper functions to ensure a correct
reading of data from multiple sources. Data can be read in one or two rows.
}
\examples{
# Loading external data requires the path (including the filename.txt)
\dontrun{
dat <- bimload(path)}

}
