% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimpos.R
\name{bimpos}
\alias{bimpos}
\title{Cutoff Determination}
\usage{
bimpos(ydata = NULL, GT = NULL, simOpt = NULL, limitToRun = 5, showPlot = TRUE)
}
\arguments{
\item{ydata}{curated data.frame from the preprocessing function.}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{limitToRun}{limit to this number of repetitions for the
randomizations (default 5, you'll probably need more)}

\item{showPlot}{plot the worth plot (default=FALSE)}
}
\value{
list with the simulation errors (simerrors) and the plot object (p).
}
\description{
The \code{bimpos} estimates a reasonable position of an incomplete item by
using an uninformed simulation of missing data.
}
