% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimposim.R
\name{bimposim}
\alias{bimposim}
\title{Integrated Position Simulation Function}
\usage{
bimposim(
  dat = NULL,
  testitem = NULL,
  GT = NULL,
  tested_1 = NULL,
  tested_2 = NULL,
  threshold = 10,
  runs = 50,
  tcut = NULL,
  seed = FALSE,
  ylim = c(0, 1)
)
}
\arguments{
\item{dat}{raw data from the bim_load function}

\item{testitem}{item of interest (to determine the hypothetical position of the item)}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{tested_1}{testitem is tested against this item}

\item{tested_2}{testitem is tested against this item as well}

\item{threshold}{the default cutoff for the lower end of the Iratio is 10 (\%).}

\item{runs}{the number of simulation runs (default is 50)}

\item{tcut}{the transitivity ratio cut (default is NULL) - at the default level
the function uses the threshold parameter as cutoff.}

\item{ylim}{y limits of the worth plot (default c(0,1))}
}
\value{
list with the Iratio cutoff recommendation, the simulated data per run,
the frequency distribution of the simulated positions with the corresponding
Iratios, the simulated position plot (a bubbleplot), the optimal run with the
best simulated transitivity as well as the reconstructed worth plot, including
the simulated item at the optimally found position
}
\description{
The \code{bimposim} function estimates the position of an item for which
limited data are available by testing it against 2 other items from the fully
available data. The open item positions are randomly filled and simulated.
The postion estimation is derived from the best simulated transitivity -
therefore, the outcome is dependent on the number of runs, relative to the
totally available item combinations and data completeness. The transitivity
cutoff is set to be no less than 10\% than the transitivity in the worth model
without the tested item. The user can, however, specify other cutoffs.
}
