% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimpre.R
\name{bimpre}
\alias{bimpre}
\title{Data Preprocessing Function}
\usage{
bimpre(
  dat = NULL,
  GT = NULL,
  simOpt = NULL,
  deviation = 0,
  minQuantity = 0,
  verbose = TRUE
)
}
\arguments{
\item{dat}{curated data.frame obtained from the import function
(unified format)}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{deviation}{how much deviation is allowed for marking ties
(default is 0 which translates to 50\%)}

\item{minQuantity}{minimum quantitiy to be recognized as difference
(default is 0)}

\item{verbose}{show function results}
}
\value{
data.frame with the raw data
}
\description{
The \code{bimpre} function is a helper function for data preprocessing.
Missing item combinations are filled and marked as ties. Simulated data are
marked and the data are prepared in the correct order (pairings) according
to the ground truth.
}
