% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimristics.R
\name{bimristics}
\alias{bimristics}
\title{bimristics - data characteristics}
\usage{
bimristics(predat = NULL, simOpt = NULL)
}
\arguments{
\item{predat}{curated data.frame obtained from the \code{bimpre} function}

\item{simOpt}{item to be checked}
}
\value{
data.frame with study characteristics
}
\description{
The \code{bimristics} function shows general data characteristics like the
number of natural ties, total number of tests, etc. It requires curated data
from the \code{bimpre} function and has to be checked for each test item
individually (simOpt).
}
