% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimsim.R
\name{bimsim}
\alias{bimsim}
\title{Preference Test Simulation with Variable Ground Truth}
\usage{
bimsim(
  ydata = NULL,
  GT = NULL,
  simOpt = NULL,
  limitToRun = 10,
  seed = TRUE,
  fval = 0.9,
  showPlot = TRUE,
  ylim = c(0, 0.7)
)
}
\arguments{
\item{ydata}{curated data.frame from the preprocessing function.}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{limitToRun}{limit to this number of repetitions for the randomizations
(default 5, you'll probably need more)}

\item{seed}{use constant seeding for stable plot results (default=TRUE)}

\item{fval}{filter value for informed transitivity checks (default = 0.9 for
90\% of all possible values), the lower the value the more informed the
positioning becomes}

\item{showPlot}{show the bubble plot with the intransitivity value vs item
position (bubble size=worth-value)}

\item{ylim}{scale y axis of the worth plot (default: c(0,0.7))}
}
\value{
list with the position frequencies and the plot object
}
\description{
The \code{bimsim} function covers the informed simulation of the preference
test data, e.g., when an additional item was tested but has ties and/or
missing item combinations according to the ground truth. With the help of
intransitivity calculations a cutoff can be set to exclude any simulation
results that do not show transitivity.
}
