% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimUninformed.R
\name{bimUninformed}
\alias{bimUninformed}
\title{Uninformed positioning and cutoff determination}
\usage{
bimUninformed(
  ydata = NULL,
  GT = NULL,
  simOpt = NULL,
  limitToRun = 5,
  seed = TRUE,
  showPlot = TRUE,
  ylim = c(-0.5, 1.5)
)
}
\arguments{
\item{ydata}{curated data.frame from the preprocessing function.}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{limitToRun}{limit to this number of repetitions for the randomizations
(default 5, you'll probably need more)}

\item{seed}{use constant seeding for stable plot results (default=TRUE)}

\item{showPlot}{show stability plot}

\item{ylim}{y-limits of the output plot (default c(-0.5,1.5))}
}
\value{
list with the randomized worth values (W), the result of the posthoc
tests (posthocs), the cutoff values (cutoffs) and
the conditional plot output (p).
}
\description{
The \code{bimUninformed} estimates reasonable randomization cutoffs for
simOption items. The cutoff is determined by monitoring the adjusted p-value
development in post-hoc ANOVA tests. The optimal cutoff is reached when
the adjusted p-values show no overlap, or zero deviation in the plot.
}
