% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimworth.R
\name{bimworth}
\alias{bimworth}
\title{Worth Calculation}
\usage{
bimworth(
  ydata = NULL,
  GT = NULL,
  simOpt = NULL,
  randOP = FALSE,
  intrans = FALSE,
  showPlot = FALSE,
  ylim = c(0, 0.7),
  verbose = FALSE
)
}
\arguments{
\item{ydata}{curated data.frame obtained from the bimpre function}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{randOP}{randomize the test result for open/undecided pairs
(e.g., data marked as ties)}

\item{intrans}{calculate intransitivity during worth value estimation
(calculation intense! TRUE/FALSE)}

\item{showPlot}{plot the worth plot (default=FALSE)}

\item{ylim}{y limits of the worth plot (default c(0,0.7))}

\item{verbose}{show model outputs (default= FALSE)}
}
\value{
list with the worth values for each item and the total number of
simulated combinations based on simOption
}
\description{
The \code{bimworth} calculates the worth model using the curated preference
test data (e.g, from bimpre).
}
