% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simsim.R
\name{simsim}
\alias{simsim}
\title{Item Position Simulation in a Preference Test}
\usage{
simsim(
  data = NULL,
  simOpt = NULL,
  GT = NULL,
  seeding = TRUE,
  runs = NULL,
  truepos = NULL,
  verbose = FALSE,
  path = NULL
)
}
\arguments{
\item{data}{rawdata from the bimload function}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{seeding}{use constant seeding for stable simulation results
(default=TRUE)}

\item{runs}{the number of simulation runs (e.g., 200)}

\item{truepos}{the true position of the item (most likely unknown); used for
performance testing}

\item{verbose}{FALSE silences the bimpre output on the number of simulated
ties}

\item{path}{path where the results of the simulations are saved as *.txt
files. If left empty (NULL), the results will only be stored in memory.}

\item{filter.crit}{the filter criterion for the plot (defaults to "CE":
consensus error; alternatively the intransitivity ratio (Iratio) can be
shown)}
}
\value{
list with the position frequencies and the plot object
}
\description{
With \code{simsim} function the putative position of new items in a
preference test can be simulated based on its Worth Value. This requires a
minimum set of tested items and the remaining combinations with the ground
truth (GT) will be simulated. The performance of the simulation can be
assessed with the amount of intransitive results (expressed as a ratio:
Iratio) and the overall consensus error (CE). The more experimentally tested
item combinations enter the simulation, the more accurate the simulated
position will be.
}
