% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZickeZacke.R
\docType{data}
\name{ZickeZacke}
\alias{ZickeZacke}
\title{ZickeZacke example data}
\format{
A data frame with 32 rows and 5 variables:
\describe{
\item{subjectID}{animal or subject ID}
\item{optionA}{First item presented to the subject}
\item{optionB}{Second item presented to the subject}
\item{quantityA}{amount of A; this can also be binary (0,1)}
\item{quantityB}{amount of B; this can also be binary (0,1)}
}
}
\source{
\url{https://talbotsr.com/simsalRbim/}
}
\usage{
ZickeZacke
}
\description{
An example data set with artificial preference test data.
The 'HoiHoiHoi' item contains 1 tie to have variance for
a simulation
}
\keyword{datasets}
