% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimsim.R
\name{bimsim}
\alias{bimsim}
\title{Preference Test Simulation with Variable Ground Truth}
\usage{
bimsim(
  rawdat = NULL,
  GT = GT,
  simOpt = simOpt,
  limitToRun = 5,
  fval = 100,
  deviation = 0,
  minQuantity = 0,
  seed = TRUE,
  showPlot = TRUE,
  ylim = c(0, 0.7)
)
}
\arguments{
\item{rawdat}{rawdata from the bimload function}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{limitToRun}{limit to this number of repetitions for the randomizations
(default 5, you'll probably need more)}

\item{fval}{filter value for consensus error (default = 100 for
100\% of all errors), the lower the value the more informed the
positioning becomes}

\item{deviation}{how much deviation is allowed for marking ties
(default is 0 which translates to 50\%)}

\item{minQuantity}{minimum quantitiy to be recognized as difference
(default is 0)}

\item{seed}{use constant seeding for stable plot results (default=TRUE)}

\item{showPlot}{show simulation runs with consensus error bubbles}

\item{ylim}{scale y axis of the worth plot (default: c(0,0.7))}
}
\value{
list with the position frequencies and the plot object
}
\description{
The \code{bimsim} function covers the informed simulation of the preference
test data, e.g., when an additional item was tested but has ties and/or
missing item combinations according to the ground truth. With the help of
intransitivity calculations a cutoff can be set to exclude any simulation
results that do not show transitivity.
}
