% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimsim.R
\name{bimsim}
\alias{bimsim}
\title{Preference Test Simulation with Variable Ground Truth}
\usage{
bimsim(
  rawdat = NULL,
  GT = GT,
  simOpt = simOpt,
  filter.crit = "CE",
  limitToRun = 5,
  tcut = 0.9,
  deviation = 0,
  minQuantity = 0,
  seed = TRUE,
  showPlot = TRUE,
  ylim = c(0, 0.7)
)
}
\arguments{
\item{rawdat}{rawdata from the bimload function}

\item{GT}{item list with the ground truth (GT; letters are case sensitive!)}

\item{simOpt}{item to be checked (this can be an item from the GT or a new
one. Remove the item from GT if specified here!)}

\item{filter.crit}{the filter criterion for the plot (defaults to "CE":
consensus error; alternatively the intransitivity ratio (Iratio) can be
shown)}

\item{limitToRun}{limit to this number of repetitions for the randomizations
(default 5, you'll probably need more)}

\item{tcut}{threshold (cutoff) for the filter.crit variable (default = 1 for
100\% of the values); i.e., when filter.crit="CE" and tcut=0.9, the plot will
show 90\% of the consensus errors. When filter.crit="Iratio" and tcut =0.9,
the plot will show items with transitivities > 90\%.}

\item{deviation}{how much deviation is allowed for marking ties
(default is 0 which translates to 50\%; deviation =5 adds 5\% to 50\% = 55\%)}

\item{minQuantity}{minimum quantity to be recognized as difference
(default is 0)}

\item{seed}{use constant seeding for stable plot results (default=TRUE)}

\item{showPlot}{show simulation runs with consensus error bubbles}

\item{ylim}{scale y axis of the worth plot (default: c(0,0.7))}
}
\value{
list with the position frequencies and the plot object
}
\description{
The \code{bimsim} function covers the informed simulation of the preference
test data, e.g., when an additional item was tested but has ties and/or
missing item combinations according to the ground truth. With the help of
intransitivity calculations a cutoff can be set to exclude any simulation
results that do not show transitivity.
}
